[Release 3.3][CP][AMD] Use AxisAnalysis for buffer op offsets to compute vector size (#6126)
This PR changes the lowering to use the `AxisAnalysis` on the offsets to
get the correct vector size for buffer operations. The lowering of
buffer operations did not analyze the contiguity of the offset values
but was only looking at the layout. This results in wrong vectorization
when the offsets are not contiguous per lane (e.g. strided loads).

It also adds a lit test for strided buffer loads and stores and adjusts
some existing buffer operation lit tests to provide enough information
for the `AxisAnalysis` to work.

(cherry picked from commit cdb53266e6c251d91a2c321d64e8466caff129a9)